Revert "Merge branch 'feat/code_snippet_integration' of https://github.com/Better-Commerce/betterstorefront into Fix-Guest-Checkout"
This reverts commit 67038a0ea6a7dc64995654c1c7994300121a0387, reversing
changes made to baf09cab58224afd43e832fb1c84d8f9e24a3321.
