+++ framework/bettercommerce/index.ts
@@ -1,5 +1,6 @@
 import { getCommerceProvider, useCommerce as useCoreCommerce } from '@commerce'
 import { betterCommerceProvider, BetterCommerceProvider } from './provider'
+import useContentSnippet from './content/use-content-snippet';
 
 export { betterCommerceProvider }
 export type { BetterCommerceProvider }
@@ -7,3 +8,4 @@ export type { BetterCommerceProvider }
 export const CommerceProvider = getCommerceProvider(betterCommerceProvider)
 
 export const useCommerce = () => useCoreCommerce<BetterCommerceProvider>()
+export { useContentSnippet };
